*** empty log message ***

+++ src/bdz_ph.h
@@ -9,7 +9,7 @@ typedef struct __bdz_ph_config_data_t bdz_ph_config_data_t;
 bdz_ph_config_data_t *bdz_ph_config_new();
 void bdz_ph_config_set_hashfuncs(cmph_config_t *mph, CMPH_HASH *hashfuncs);
 void bdz_ph_config_destroy(cmph_config_t *mph);
-cmph_t *bdz_ph_new(cmph_config_t *mph, float c);
+cmph_t *bdz_ph_new(cmph_config_t *mph, double c);
 
 void bdz_ph_load(FILE *f, cmph_t *mphf);
 int bdz_ph_dump(cmph_t *mphf, FILE *f);
